[ doc ] add two poll examples

+++ posix/support/posix.c
@@ -251,11 +251,11 @@ uint32_t get_pollfd_fd(struct pollfd *fd) {
   return fd->fd;
 }
 
-uint32_t get_pollfd_events(struct pollfd *fd) {
+uint16_t get_pollfd_events(struct pollfd *fd) {
   return fd->events;
 }
 
-uint32_t get_pollfd_revents(struct pollfd *fd) {
+uint16_t get_pollfd_revents(struct pollfd *fd) {
   return fd->revents;
 }
 
@@ -263,11 +263,11 @@ void set_pollfd_fd(struct pollfd *fd, uint32_t val) {
   fd->fd = val;
 }
 
-void set_pollfd_events(struct pollfd *fd, uint32_t val) {
+void set_pollfd_events(struct pollfd *fd, uint16_t val) {
   fd->events = val;
 }
 
-void set_pollfd_revents(struct pollfd *fd, uint32_t val) {
+void set_pollfd_revents(struct pollfd *fd, uint16_t val) {
   fd->revents = val;
 }
